# public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "soca web api v1.0",
        "description": "soca web API Description",
        "version": "1.0"
    },
    "paths": {
...
        "/api/Users/{pg}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get users data",
                "operationId": "9c3d8be2cb3110b2796023b4220b2ae1",
                "parameters": [
                    {
                        "name": "pg",
                        "in": "path",
                        "description": "Page number for pagination",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of USERS"
                    }
                }
            }
        },
        "/api/Users/Previous/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get users data",
                "operationId": "e961a38aafb7c326ef517e4490b33051",
                "parameters": [
                    {
                        "name": "pg",
                        "in": "path",
                        "description": "Page number for pagination",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of USERS"
                    }
                }
            }
        },
        "/api/Users/Next/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get users data",
                "operationId": "5d5e117ca158303ad02af403319415de",
                "parameters": [
                    {
                        "name": "pg",
                        "in": "path",
                        "description": "Page number for pagination",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of USERS"
                    }
                }
            }
        },
        "/api/Users/{id}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete users data",
                "operationId": "e5bf9edc5ed1ddef1f8ce267a8f3c4ac",
                "parameters": [
                    {
                        "name": "pg",
                        "in": "path",
                        "description": "Page number for pagination",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Users Id"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "update users data",
                "operationId": "8c8d11fc255402a33f97543ebfd357b2",
                "parameters": [
                    {
                        "name": "pg",
                        "in": "path",
                        "description": "Page number for pagination",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of USERS"
                    }
                }
            }
        },
        "/api/GetReaderFromGIdx/{GIdxStr}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get Reader data",
                "operationId": "e4af60ef32c0e4d939d99fe498aa8334",
                "parameters": [
                    {
                        "name": "GIdxStr",
                        "in": "path",
                        "description": "獲取群組的讀卡機清單資料",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of USERS"
                    }
                }
            }
        },
        "/api/Users/": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create users data",
                "operationId": "03d7d29f981dc150ea0ef3af5a9ba6be",
                "parameters": [
                    {
                        "name": "pg",
                        "in": "path",
                        "description": "Page number for pagination",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Users Id"
                    }
                }
            }
        }
    },
...
}